requirements.txt
streamlit
pandas
numpy
plotly
streamlit
ollama
streamlit
groq
streamlit
groq
qrcode[pil]
Pillow      
twilio
razorpay